moved counter and link to shop cart
