+++ src/app.jsx
@@ -4,6 +4,7 @@ import ItemDetailsContainer from "./components/containers/item-details/item-deta
 import {BrowserRouter, Routes, Route, Navigate} from "react-router-dom";
 import Home from "./components/presentationals/home/home";
 import NavBarContainer from "./components/containers/nav-bar/nav-bar.container";
+import Cart from "./components/presentationals/cart/cart";
 
 function App() {
   return (
@@ -11,6 +12,7 @@ function App() {
         <NavBarContainer />
         <Routes>
             <Route path="/" element={<Home />}/>
+            <Route path="/cart" element={<Cart />}/>
             <Route path="/products" element={<ItemListContainer />}/>
             <Route path="/product/:id" element={<ItemDetailsContainer />}/>
             <Route path="/products/category/:id" element={<ItemListContainer />}/>
