+++ src/components/presentationals/cart/cart.jsx
@@ -1,0 +1,9 @@
+const Cart = () => {
+    return(
+        <div className="d-flex flex-column align-items-center">
+            <h2>El Carrito</h2>
+        </div>
+    )
+}
+
+export default Cart;
