+++ src/components/presentationals/item-details/item-details.jsx
@@ -1,4 +1,6 @@
 import ItemCounterContainer from "../../containers/item-counter/item-counter.container";
+import {NavLink} from "react-router-dom";
+import {Button} from "react-bootstrap";
 
 const ItemDetail = (props) => {
     return (
@@ -10,6 +12,9 @@ const ItemDetail = (props) => {
                 <h2>Nombre: {props.product.name}</h2>
                 <h4>Precio: {props.product.price}</h4>
                 <ItemCounterContainer stock={props.product.stock}/>
+                <NavLink to={"/cart"} className="link-no-decoration">
+                    <Button variant="primary">Ir al carrito</Button>
+                </NavLink>
             </div>
         </div>
     )
